Fix a bunch of warnings

+++ 3rdparty/gameplay/gameplay/src/MeshBatch.cpp
@@ -45,14 +45,14 @@ namespace gameplay
                 // Create a degenerate triangle to connect separate triangle strips
                 // by duplicating the previous and next vertices.
                 m_indices.emplace_back(m_indices.back());
-                m_indices.emplace_back(currentVertexCount);
+                m_indices.emplace_back(gsl::narrow<uint16_t>(currentVertexCount));
             }
 
             // Loop through all indices and insert them, with their values offset by
             // 'vertexCount' so that they are relative to the first newly inserted vertex.
             for( size_t i = 0; i < indices.size(); ++i )
             {
-                m_indices.emplace_back(indices[i] + currentVertexCount);
+                m_indices.emplace_back(gsl::narrow<uint16_t>(indices[i] + currentVertexCount));
             }
         }
