+++ 3rdparty/gameplay/gameplay/src/ScreenOverlay.cpp
@@ -47,7 +47,7 @@ namespace gameplay
             BOOST_THROW_EXCEPTION(std::runtime_error("Cannot create screen overlay because the viewport is empty"));
         }
 
-        _image = std::make_shared<Image>(vp.width, vp.height);
+        _image = std::make_shared<Image>(static_cast<GLint>(vp.width), static_cast<GLint>(vp.height));
         _image->fill({0,0,0,0});
 
         _texture = std::make_shared<gl::Texture>(GL_TEXTURE_2D);
